Relative dependency injection
