+++ scripts/deploy/inject_dependencies.py
@@ -1,0 +1,405 @@
+#!/usr/bin/env python3
+"""
+Inject dependencies into AWS example projects.
+
+This script:
+1. Injects backend.tf files for Terraform state management
+2. Modifies main.tf files to use local module sources instead of registry modules
+3. Can revert main.tf changes back to registry modules
+
+The modifications are intended for local development and CI/CD use.
+"""
+
+import os
+import re
+import yaml
+import argparse
+from pathlib import Path
+from typing import List, Dict, Set, Tuple
+
+
+def load_config(config_path: str) -> Dict:
+    """Load the integration test configuration YAML file."""
+    with open(config_path, 'r') as f:
+        return yaml.safe_load(f)
+
+
+def get_aws_projects(config: Dict) -> Set[tuple]:
+    """
+    Extract AWS project paths from the configuration.
+    
+    Returns a set of tuples: (project_path, deploy_dir, project_type)
+    """
+    aws_projects = set()
+    
+    # Include deployment base infrastructure
+    if 'deployment_base' in config:
+        for project in config['deployment_base']:
+            test_type = project.get('type', '')
+            if test_type in ['aws-serverless', 'aws-containerized']:
+                path = project.get('path', '')
+                deploy_dir = project.get('deploy_dir', 'deploy')
+                if path:
+                    aws_projects.add((path, deploy_dir, test_type))
+    
+    # Process both nightly and weekly test configs
+    for schedule in ['nightly', 'weekly']:
+        if schedule in config and 'tests' in config[schedule]:
+            for test in config[schedule]['tests']:
+                test_type = test.get('type', '')
+                if test_type in ['aws-serverless', 'aws-containerized']:
+                    path = test.get('path', '')
+                    deploy_dir = test.get('deploy_dir', 'deploy')
+                    if path:
+                        aws_projects.add((path, deploy_dir, test_type))
+    
+    return aws_projects
+
+
+def generate_backend_tf(template_content: str, project_path: str, project_type: str) -> str:
+    """
+    Generate backend.tf content from template with project-specific values.
+    
+    Args:
+        template_content: The template file content
+        project_path: Path to the project (e.g., examples/aws-serverless/adk)
+        project_type: Type of project (aws-serverless or aws-containerized)
+    
+    Returns:
+        Generated backend.tf content
+    """
+    # Extract project name from path (e.g., "adk" from "examples/aws-serverless/adk")
+    project_name = project_path.split('/')[-1]
+    
+    # Generate state key based on project path
+    state_key = f"{project_path}/terraform.tfstate"
+    
+    # Use placeholder values - these should be configured per environment
+    bucket_name = "agent-kernel-terraform-state-bucket"
+    dynamodb_table = "ak-terraform-state-lock"
+    region = "ap-southeast-2"
+    
+    # Replace placeholders in template
+    content = template_content.replace("{bucket_name}", bucket_name)
+    content = content.replace("{state_key}", state_key)
+    content = content.replace("{region}", region)
+    content = content.replace("{dynamodb_table}", dynamodb_table)
+    
+    return content
+
+
+def calculate_relative_path(from_path: str, project_type: str) -> str:
+    """
+    Calculate relative path from example deploy directory to module source.
+    
+    Args:
+        from_path: Path to the example (e.g., examples/aws-serverless/adk)
+        project_type: Type of project (aws-serverless or aws-containerized)
+    
+    Returns:
+        Relative path to the module source
+    """
+    # Module is at ak-deployment/ak-aws/serverless or ak-deployment/ak-aws/containerized
+    if project_type == 'aws-serverless':
+        module_subdir = 'serverless'
+    elif project_type == 'aws-containerized':
+        module_subdir = 'containerized'
+    else:
+        return None
+    
+    # Calculate depth from deploy dir (e.g., examples/aws-serverless/adk/deploy)
+    # Need to go up to workspace root, then down to module
+    depth = len(from_path.split('/')) + 1  # +1 for deploy dir
+    up_levels = '../' * depth
+    
+    return f"{up_levels}ak-deployment/ak-aws/{module_subdir}"
+
+
+def modify_main_tf(main_tf_path: str, project_path: str, project_type: str) -> bool:
+    """
+    Modify main.tf to use local module source instead of registry.
+    
+    Args:
+        main_tf_path: Path to main.tf file
+        project_path: Path to the project (e.g., examples/aws-serverless/adk)
+        project_type: Type of project (aws-serverless or aws-containerized)
+    
+    Returns:
+        True if modifications were made, False otherwise
+    """
+    if not os.path.exists(main_tf_path):
+        return False
+    
+    with open(main_tf_path, 'r') as f:
+        content = f.read()
+    
+    original_content = content
+    
+    # Determine which registry module to replace
+    if project_type == 'aws-serverless':
+        registry_source = 'yaalalabs/ak-serverless/aws'
+    elif project_type == 'aws-containerized':
+        registry_source = 'yaalalabs/ak-containerized/aws'
+    else:
+        return False
+    
+    # Check if already using local source
+    if registry_source not in content:
+        return False
+    
+    # Calculate relative path
+    relative_path = calculate_relative_path(project_path, project_type)
+    if not relative_path:
+        return False
+    
+    # Replace source line
+    # Pattern: source = "yaalalabs/ak-serverless/aws" or source = "yaalalabs/ak-containerized/aws"
+    source_pattern = rf'(\s+source\s*=\s*"){re.escape(registry_source)}(")'
+    content = re.sub(source_pattern, rf'\1{relative_path}\2', content)
+    
+    # Comment out version line (should be on next line after source)
+    # Pattern: version = "x.x.x"
+    version_pattern = r'(\s+)(version\s*=\s*"[^"]*")'
+    content = re.sub(version_pattern, r'\1# \2  # Commented for local development', content)
+    
+    # Only write if changes were made
+    if content != original_content:
+        with open(main_tf_path, 'w') as f:
+            f.write(content)
+        return True
+    
+    return False
+
+
+def revert_main_tf(main_tf_path: str, project_path: str, project_type: str) -> bool:
+    """
+    Revert main.tf back to use registry module source.
+    
+    Args:
+        main_tf_path: Path to main.tf file
+        project_path: Path to the project (e.g., examples/aws-serverless/adk)
+        project_type: Type of project (aws-serverless or aws-containerized)
+    
+    Returns:
+        True if modifications were made, False otherwise
+    """
+    if not os.path.exists(main_tf_path):
+        return False
+    
+    with open(main_tf_path, 'r') as f:
+        content = f.read()
+    
+    original_content = content
+    
+    # Determine which registry module to restore
+    if project_type == 'aws-serverless':
+        registry_source = 'yaalalabs/ak-serverless/aws'
+    elif project_type == 'aws-containerized':
+        registry_source = 'yaalalabs/ak-containerized/aws'
+    else:
+        return False
+    
+    # Calculate relative path to check if it's currently using local source
+    relative_path = calculate_relative_path(project_path, project_type)
+    if not relative_path:
+        return False
+    
+    escaped_relative_path = re.escape(relative_path)
+    
+    # Check if currently using local source
+    if relative_path not in content:
+        return False
+    
+    # Replace local source with registry source
+    # Pattern: source = "../../../../ak-deployment/ak-aws/serverless" or similar
+    source_pattern = rf'(\s+source\s*=\s*"){escaped_relative_path}(")'
+    content = re.sub(source_pattern, rf'\1{registry_source}\2', content)
+    
+    # Uncomment version line
+    # Pattern: # version = "x.x.x"  # Commented for local development
+    version_pattern = r'(\s+)# (version\s*=\s*"[^"]*")\s*# Commented for local development'
+    content = re.sub(version_pattern, r'\1\2', content)
+    
+    # Only write if changes were made
+    if content != original_content:
+        with open(main_tf_path, 'w') as f:
+            f.write(content)
+        return True
+    
+    return False
+
+
+def revert_dependencies(
+    workspace_root: str,
+    aws_projects: Set[Tuple[str, str, str]]
+) -> None:
+    """
+    Revert main.tf files back to registry modules.
+    
+    Args:
+        workspace_root: Root directory of the workspace
+        aws_projects: Set of AWS project tuples (path, deploy_dir, type)
+    """
+    module_count = 0
+    
+    for project_path, deploy_dir, project_type in sorted(aws_projects):
+        full_project_path = os.path.join(workspace_root, project_path)
+        deploy_path = os.path.join(full_project_path, deploy_dir)
+        
+        if not os.path.exists(deploy_path):
+            print(f"⚠️  Warning: Deploy directory not found: {deploy_path}")
+            continue
+        
+        # Revert main.tf to use registry modules
+        main_tf_path = os.path.join(deploy_path, "main.tf")
+        if revert_main_tf(main_tf_path, project_path, project_type):
+            print(f"✅ Reverted main.tf -> {main_tf_path}")
+            module_count += 1
+    
+    print(f"\n✨ Successfully reverted main.tf in {module_count} projects")
+
+
+def inject_dependencies(
+    workspace_root: str,
+    template_path: str,
+    aws_projects: Set[Tuple[str, str, str]]
+) -> None:
+    """
+    Inject backend.tf and modify main.tf in AWS example projects.
+    
+    Args:
+        workspace_root: Root directory of the workspace
+        template_path: Path to backend.tf.template
+        aws_projects: Set of AWS project tuples (path, deploy_dir, type)
+    """
+    # Read template
+    with open(template_path, 'r') as f:
+        template_content = f.read()
+    
+    backend_count = 0
+    module_count = 0
+    
+    for project_path, deploy_dir, project_type in sorted(aws_projects):
+        full_project_path = os.path.join(workspace_root, project_path)
+        deploy_path = os.path.join(full_project_path, deploy_dir)
+        
+        if not os.path.exists(deploy_path):
+            print(f"⚠️  Warning: Deploy directory not found: {deploy_path}")
+            continue
+        
+        # 1. Generate and write backend.tf
+        backend_tf_path = os.path.join(deploy_path, "backend.tf")
+        backend_content = generate_backend_tf(template_content, project_path, project_type)
+        
+        with open(backend_tf_path, 'w') as f:
+            f.write(backend_content)
+        print(f"✅ Injected backend.tf -> {backend_tf_path}")
+        backend_count += 1
+        
+        # 2. Modify main.tf to use local modules
+        main_tf_path = os.path.join(deploy_path, "main.tf")
+        if modify_main_tf(main_tf_path, project_path, project_type):
+            print(f"✅ Modified main.tf -> {main_tf_path}")
+            module_count += 1
+    
+    print(f"\n✨ Successfully injected backend.tf into {backend_count} projects")
+    print(f"✨ Successfully modified main.tf in {module_count} projects")
+
+
+def inject_files(
+    workspace_root: str,
+    template_path: str,
+    aws_projects: Set[tuple]
+) -> None:
+    """
+    Inject backend.tf files into AWS example projects.
+    
+    Args:
+        workspace_root: Root directory of the workspace
+        template_path: Path to backend.tf.template
+        aws_projects: Set of AWS project tuples (path, deploy_dir, type)
+    """
+    # Read template
+    with open(template_path, 'r') as f:
+        template_content = f.read()
+    
+    injected_count = 0
+    
+    for project_path, deploy_dir, project_type in sorted(aws_projects):
+        full_project_path = os.path.join(workspace_root, project_path)
+        deploy_path = os.path.join(full_project_path, deploy_dir)
+        
+        if not os.path.exists(deploy_path):
+            print(f"⚠️  Warning: Deploy directory not found: {deploy_path}")
+            continue
+        
+        # Generate and write backend.tf
+        backend_tf_path = os.path.join(deploy_path, "backend.tf")
+        backend_content = generate_backend_tf(template_content, project_path, project_type)
+        
+        with open(backend_tf_path, 'w') as f:
+            f.write(backend_content)
+        print(f"✅ Injected backend.tf -> {backend_tf_path}")
+        
+        injected_count += 1
+    
+    print(f"\n✨ Successfully injected files into {injected_count} projects")
+
+
+def main():
+    """Main entry point."""
+    parser = argparse.ArgumentParser(
+        description='Inject or revert dependencies in AWS example projects'
+    )
+    parser.add_argument(
+        '--revert',
+        action='store_true',
+        help='Revert main.tf files back to registry modules'
+    )
+    
+    args = parser.parse_args()
+    
+    # Determine workspace root (assuming script is in scripts/deploy)
+    script_dir = Path(__file__).parent
+    workspace_root = script_dir.parent.parent
+    
+    # Paths to template and config
+    template_path = script_dir / "backend.tf.template"
+    config_path = workspace_root / ".github" / "integration-test-config.yaml"
+    
+    if not config_path.exists():
+        print(f"❌ Error: Config file not found: {config_path}")
+        return 1
+    
+    print("🔍 Loading integration test configuration...")
+    config = load_config(str(config_path))
+    
+    print("🔍 Identifying AWS projects...")
+    aws_projects = get_aws_projects(config)
+    print(f"   Found {len(aws_projects)} AWS projects")
+    
+    if args.revert:
+        # Revert mode: restore registry modules
+        print("\n🔄 Reverting to registry modules...")
+        revert_dependencies(
+            str(workspace_root),
+            aws_projects
+        )
+    else:
+        # Inject mode: inject backend.tf and modify main.tf
+        if not template_path.exists():
+            print(f"❌ Error: Template not found: {template_path}")
+            return 1
+        
+        print("\n📝 Injecting dependencies...")
+        inject_dependencies(
+            str(workspace_root),
+            str(template_path),
+            aws_projects
+        )
+    
+    return 0
+
+
+if __name__ == "__main__":
+    exit(main())
